UI makes more sense now, especially when a new project has no sprints
