+++ client/src/components/SprintList/AddSprintButton/index.js
@@ -1,0 +1,14 @@
+import React from 'react'
+import './style.css'
+
+
+
+const CreateSprintButton = ({ openAddSprintModal }) => {
+  return (
+    <div id="add-sprint" onClick={() => openAddSprintModal()}>
+      +
+    </div>
+  )
+}
+
+export default CreateSprintButton
