+++ client/src/components/SprintList/SprintListEmpty/index.js
@@ -1,0 +1,15 @@
+import React from 'react'
+
+
+
+const SprintListEmpty = () => {
+  return (
+    <div className="sprintlist-wrapper">
+      <div className="sprintlist-container">
+        Currently no sprints
+      </div>
+    </div>
+  )
+}
+
+export default SprintListEmpty
